detection_engine/requirements.txt
kafka-python
redis
psycopg2-binary